src/python/functions/func_trigger/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
xmltodict
azure-storage-blob
azure-core
opencensus-extension-azure-functions
